+++ .pre-commit-config.yaml
@@ -21,8 +21,8 @@ repos:
     rev: 1.13.0
     hooks:
       - id: blacken-docs
-        additional_dependencies: [black==22.10.0]
-        args: [-l, '77', -t, py39]
+        additional_dependencies: [black==23.1.0]
+        args: [-l, '76', -t, py39]
 
   - repo: https://github.com/PyCQA/flake8
     rev: 6.0.0
