+++ docs/user-guide/index.rst
@@ -27,4 +27,4 @@ User guide
    fastapi-errors
    datetime
    uvicorn
-   slack
+   slack-webhook
